chore; move common components to shared
