+++ libs/chat-shared/src/components/MarkdownRenderer/Table/MarkdownTable.module.scss
