+++ libs/chat-shared/src/hooks/useStreamedMarkdownContent.ts
