+++ libs/chat-shared/src/hooks/useTableScroll.ts
