+++ libs/chat-shared/src/index.ts
@@ -7,6 +7,7 @@ export * from './models/deployment-configuration';
 export * from './models/deployment';
 export * from './types/attachment';
 export * from './types/mime-type';
+export * from './types/code-editor';
 export * from './utils/merge-class';
 export * from './utils/build-css-vars';
 export * from './utils/message';
@@ -19,3 +20,7 @@ export * from './constants/dial';
 export * from './components/DeploymentIcon/DeploymentIcon';
 export * from './components/Highlight/Highlight';
 export * from './components/PanelEmptyState/PanelEmptyState';
+export * from './components/MarkdownRenderer/MarkdownRenderer';
+export * from './components/MarkdownRenderer/MDMessageViewer';
+export * from './components/MarkdownRenderer/CodeBlock/CodeBlock';
+export * from './components/MarkdownRenderer/Table/MarkdownTable';
