+++ libs/chat-shared/src/types/code-editor.ts
@@ -1,0 +1,5 @@
+/** Color theme for syntax highlighting. */
+export enum CodeBlockTheme {
+  Dark = 'dark',
+  Light = 'light',
+}
