+++ libs/conversation-messages/src/index.ts
@@ -1,16 +1,3 @@
-export {
-  MarkdownRenderer,
-  defaultMarkdownComponents,
-} from './components/Markdown/MarkdownRenderer';
-export type {
-  MarkdownRendererClassNames,
-  MarkdownRendererProps,
-} from './components/Markdown/MarkdownRenderer';
-export {
-  CodeBlockTheme,
-  MarkdownCodeBlock,
-} from './components/Markdown/MarkdownCodeBlock';
-export type { MarkdownCodeBlockProps } from './components/Markdown/MarkdownCodeBlock';
 export { MessageBubble } from './components/MessageBubble/MessageBubble';
 export { UserMessageBubble } from './components/MessageBubble/UserMessageBubble';
 export { AssistantMessageBubble } from './components/MessageBubble/AssistantMessageBubble';
